% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{SeqFrame}
\alias{SeqFrame}
\title{SeqFrame}
\usage{
SeqFrame(df=NULL,keyword=NULL,file=character(0),con=character(0),annotation=data.frame(),)
}
\arguments{
\item{df}{data.frame to be converted to SeqFrame. if data.frame have five columns, i.e. "chr", "start","end","strand","symbol" it will put into annotation slot, one can use annotation() method to access the corresponding genomic annotation, else it is simply converted to seqFrame with empty annotation, for simple visulization. then it needs to be all numbers.}

\item{annotation}{a data.frame containing annotation information, need to have at least five columns i.e. "chr", "start","end","strand","symbol".}

\item{keyword}{a list including descriptions for the columns. see keyword in flowFrame for detail.}

\item{file}{comma separated file, tab delimited file. Header is required. if data have five columns, i.e. "chr", "start","end","strand","symbol" it will put into annotation slot, else it is simply converted to seqFrame with empty annotation, for simple visulization. then it needs to be all numbers.}

\item{con}{filename to file formats representing annotated genomic intervals, i.e. gff, bed, bed15, bedGraph, wig and bigWig format. File type is automatically detected by file extension.}
}
\value{
A SeqFrame containing data and annotation. Users can use all flowCore and flowViz visulization functions to subset, split data based on its geometric property, and use annotation to access the corresponding annotation.
}
\description{
method for coerce spread sheet like data into SeqFrame. data type include data.fame, comma seperated file, tab delimited file and file formats representing annotated genomic intervals, i.e. gff, bed, bed15, bedGraph, wig and bigWig format. Also include a facility for coercing flowFrame into SeqFrame.
}
\details{
SeqFrame is the anolog of flowFrame in flowCore. It provide an accessor for spread sheet like genomic data to be viewed as events in flow cytometry.
}
\examples{
file="/Users/shengliu/DoScience/DoScience/Projects/Clover/oocytes methylation /DNAme/NGO-GVO-methylation.csv"
SeqFrame(file=file)
}

